+++ packages/generate-en10168-pdf-template/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.7 (2024-03-01)
+
+This was a version bump only for generate-en10168-pdf-template to align it with other projects, there were no code changes.
+
 ## 0.7.6 (2024-01-17)
 
 This was a version bump only for generate-en10168-pdf-template to align it with other projects, there were no code changes.